requirements.txt
fastapi
uvicorn
pydantic
PyYAML
opencv-python
numpy
Pillow
rapidocr_onnxruntime
deep-translator
